+++ packages/webview_flutter/webview_flutter/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.4.0
+
+* Adds support to register a callback to receive JavaScript console messages. See `WebViewController.setConsoleLogCallback`.
+
+## 4.3.0
+
+* Adds support to retrieve the user agent. See `WebViewController.getUserAgent`.
+
 ## 4.2.4
 
 * Adds pub topics to package metadata.